:truck: Rust Core: split map generator module into sub-modules
